+++ scripts/build-json/compose-attributes.js
