+++ scripts/build-json/resolve-bcd.js
