+++ scripts/build-json/resolve-contributors.js
